feat: stream-b03-common-wire-serialization-parity-003 - Preserve retention accounting and close the wire-parity release gate
